Move setupNestErrorHandler to nest SDK
