+++ packages/nestjs/src/index.ts
@@ -1,5 +1,6 @@
 export * from '@sentry/node';
 
+export { setupNestErrorHandler } from './setup';
 export { init } from './sdk';
 
 export { SentryTraced } from './span-decorator';
